check permission for dish endpoints

+++ src/main/java/ua/com/obox/dbschema/dish/DishService.java
@@ -7,8 +7,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import ua.com.obox.authserver.user.User;
-import ua.com.obox.authserver.user.UserRepository;
+import ua.com.obox.authserver.user.UserService;
 import ua.com.obox.dbschema.attachment.Attachment;
 import ua.com.obox.dbschema.attachment.AttachmentRepository;
 import ua.com.obox.dbschema.category.Category;
@@ -16,6 +15,7 @@ import ua.com.obox.dbschema.category.CategoryRepository;
 import ua.com.obox.dbschema.tools.FieldUpdateFunction;
 import ua.com.obox.dbschema.tools.Validator;
 import ua.com.obox.dbschema.tools.attachment.AttachmentTools;
+import ua.com.obox.dbschema.tools.attachment.ReferenceType;
 import ua.com.obox.dbschema.tools.configuration.ValidationConfiguration;
 import ua.com.obox.dbschema.tools.exception.ExceptionTools;
 import ua.com.obox.dbschema.tools.response.BadFieldsResponse;
@@ -41,7 +41,7 @@ public class DishService {
     private final AttachmentRepository attachmentRepository;
     private final TranslationRepository translationRepository;
     private final UpdateServiceHelper serviceHelper;
-    private final UserRepository userRepository;
+    private final UserService userService;
     private final ResourceBundle translationContent = ResourceBundle.getBundle("translation.messages");
     @Value("${application.image-dns}")
     private String attachmentsDns;
@@ -50,6 +50,8 @@ public class DishService {
         String finalAcceptLanguage = CheckHeader.checkHeaderLanguage(acceptLanguage);
 
         Dish dish = dishRepository.findByDishId(dishId).orElseThrow(() -> ExceptionTools.notFoundException(".dishNotFound", finalAcceptLanguage, dishId));
+        userService.checkPermissionForUser(ReferenceType.dish, dishId, finalAcceptLanguage);
+
         System.out.println(dish.getDishId());
         System.out.println(dish.getTranslationId());
         Translation translation = translationRepository.findAllByTranslationId(dish.getTranslationId())
@@ -82,6 +84,7 @@ public class DishService {
 
     public DishResponseId createDish(Dish request, String acceptLanguage) throws JsonProcessingException {
         String finalAcceptLanguage = CheckHeader.checkHeaderLanguage(acceptLanguage);
+        userService.checkPermissionForUser(ReferenceType.category, request.getCategoryId(), finalAcceptLanguage);
         Map<String, String> fieldErrors = new ResponseErrorMap<>();
         Optional<Category> category = categoryRepository.findByCategoryId(request.getCategoryId());
         if (category.isEmpty())
@@ -120,6 +123,8 @@ public class DishService {
         Translation translation = translationRepository.findAllByTranslationId(dish.getTranslationId())
                 .orElseThrow(() -> ExceptionTools.notFoundException(".translationNotFound", finalAcceptLanguage, dishId));
 
+        userService.checkPermissionForUser(ReferenceType.dish, dishId, finalAcceptLanguage);
+
         if (request.getCategoryId() != null && !request.getCategoryId().isEmpty()) {
             Optional<Category> category = categoryRepository.findByCategoryId(request.getCategoryId());
             if (category.isEmpty())
@@ -138,6 +143,7 @@ public class DishService {
     public void deleteDishById(String dishId, String acceptLanguage) {
         String finalAcceptLanguage = CheckHeader.checkHeaderLanguage(acceptLanguage);
         Dish dish = dishRepository.findByDishId(dishId).orElseThrow(() -> ExceptionTools.notFoundException(".dishNotFound", finalAcceptLanguage, dishId));
+        userService.checkPermissionForUser(ReferenceType.dish, dishId, finalAcceptLanguage);
         dishRepository.delete(dish);
     }
 
@@ -146,6 +152,7 @@ public class DishService {
         Map<String, String> fieldErrors = new ResponseErrorMap<>();
 
         Dish dish = dishRepository.findByDishId(dishId).orElseThrow(() -> ExceptionTools.notFoundException(".dishNotFound", finalAcceptLanguage, dishId));
+        userService.checkPermissionForUser(ReferenceType.dish, dishId, finalAcceptLanguage);
 
         Optional<Attachment> attachment = attachmentRepository.findByAttachmentId(request.getImage());
         if (attachment.isEmpty())
@@ -208,7 +215,9 @@ public class DishService {
         if (value != null || required) {
             if (Objects.equals(fieldName, "name") && dish.getCategory() != null) {
                 ExistEntity<DishTranslationEntry> existEntity = new ExistEntity<>(translationRepository);
-                /*validate exists on menu level, because dish will to move from one category to other*/
+                /**
+                 * validate exists on menu level, because dish will to move from one category to other
+                 * */
                 List<Category> sameParentCategories = categoryRepository.findAllByMenu_MenuId(dish.getCategory().getMenu().getMenuId());
                 List<Dish> allDishesForMenu = new ArrayList<>();
                 for (Category category : sameParentCategories) {
@@ -243,19 +252,4 @@ public class DishService {
         translation.setContent(objectMapper.writeValueAsString(content));
         translation.setUpdatedAt(Instant.now().getEpochSecond());
     }
-
-    public boolean isDishInTenant(String dishId, String userId) {
-        Optional<Dish> optionalDish = dishRepository.findByDishId(dishId);
-        Optional<User> optionalUser = userRepository.findByUserId(userId);
-
-        if (optionalDish.isPresent() && optionalUser.isPresent()) {
-            Dish dish = optionalDish.get();
-            User user = optionalUser.get();
-            String dishTenantId = dish.getCategory().getMenu().getRestaurant().getTenant().getTenantId();
-            return dishTenantId.equals(user.getTenant().getTenantId());
-        }
-
-        return false;
-    }
-
 }
